services/email-processor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "email-processor"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "database-models",
    "google-api-python-client==2.185.0",
    "google-auth==2.41.1",
    "google-auth-oauthlib==1.2.2",
    "google-auth-httplib2==0.2.0",
    "python-dotenv==1.2.1",
    "click==8.1.8",
    "google-genai==1.46.0",
    "httpx==0.28.1",
    "structlog==25.5.0",
    "pydantic==2.12.3",
    "opentelemetry-api==1.25.0",
    "opentelemetry-sdk==1.25.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.25.0",
    "opentelemetry-instrumentation-requests==0.46b0",
    "opentelemetry-instrumentation-logging==0.46b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "depcheck",
]

[tool.uv.sources]
database-models = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["clients", "models", "pipeline", "services"]